requirements.txt
streamlit>=1.40.0
pandas>=2.0.0
pymongo>=4.6.0
certifi>=2024.0.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
pytest>=8.0.0

# Optional: legacy SQLite dev (not used when MONGO_URI is set)
sqlalchemy>=2.0.0
